requirements.txt
numpy~=1.24.1
opencv-python~=4.8.1.78
blinker~=1.6.2
pillow~=9.3.0
ultralytics~=8.0.120
requests~=2.28.1
torch~=2.0.1+cu118
torchvision~=0.15.2+cu118
matplotlib~=3.8.0
clip~=1.0
setuptools~=68.0.0
gradio~=3.35.2